% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{annotate_genes}
\alias{annotate_genes}
\title{Annotate Gene Table with Metadata}
\usage{
annotate_genes(
  gene_ids,
  key = "ENSEMBL",
  org_db = org.Hs.eg.db::org.Hs.eg.db,
  cols = c("ENSEMBL", "SYMBOL", "MAP", "GENENAME", "ENTREZID")
)
}
\arguments{
\item{gene_ids}{A vector with gene IDs}

\item{key}{The keytype (e.g. \code{ENSEMBL} or \code{SYMBOL})}

\item{org_db}{The organism annotation database to use (e.g., \code{org.Hs.eg.db}, \code{org.Mm.eg.db}).}

\item{cols}{The column values to request from the org database}
}
\value{
A data frame with gene annotations, including Ensembl ID, gene symbol, chromosomal location, gene name, and Entrez ID.
}
\description{
This function retrieves gene annotations from a specified organism database based on gene IDs.
}
