% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{create_col_scale}
\alias{create_col_scale}
\title{Create a Custom Color or Fill Scale for ggplot2}
\usage{
create_col_scale(data, column_name, custom_colors, scale_type = "color")
}
\arguments{
\item{data}{A data frame containing the column to be factored.}

\item{column_name}{A string specifying the column name to be factored.}

\item{custom_colors}{A vector of colors to be assigned to the factor levels.
The number of colors must be at least as many as the unique levels.}

\item{scale_type}{A string specifying the type of scale to return.
Accepts \code{"color"}, \code{"colour"}, or \code{"fill"}.}
}
\value{
A \code{scale_colour_manual()} or \code{scale_fill_manual()} object for ggplot2.
}
\description{
This function converts a specified column in a data frame to a factor, assigns custom colors
to its unique levels, and returns either a \code{scale_colour_manual()} or \code{scale_fill_manual()}
object for use in ggplot2.
}
\examples{
library(ggplot2)
df <- data.frame(Diff = c("Low", "Medium", "High"))
custom_colors <- c("#E5614CFF", "#8C57A2FF", "#358DB9FF")
col_scale <- create_col_scale(df, "Diff", custom_colors, "color")
ggplot(df, aes(x = Diff, y = 1, color = Diff)) +
  geom_point() +
  col_scale

fill_scale <- create_col_scale(df, "Diff", custom_colors, "fill")
ggplot(df, aes(x = Diff, y = 1, fill = Diff)) +
  geom_bar(stat = "identity") +
  fill_scale
}
