% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{rmarkdown_header}
\alias{rmarkdown_header}
\title{Print out template header for Rmarkdown}
\usage{
rmarkdown_header()
}
\description{
This copies the header of a Rmarkdown to the system clipboard (for paste later)
}
\details{
return Text
}
