% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{volcano_maker}
\alias{volcano_maker}
\title{Create a Volcano Plot}
\usage{
volcano_maker(
  df,
  title = "Volcano Plot",
  pvalue = "pvalue",
  padj = "padj",
  logFC = "log2FoldChange",
  gene_list = "",
  padj_cutoff = 0.05,
  logFC_cutoff = c(-1, 1),
  basemean_cutoff = 500,
  top_n_genes = 10,
  bottom_n_genes = 10
)
}
\arguments{
\item{df}{A data frame containing differential expression results.}

\item{title}{The title of the plot (default: "Volcano Plot").}

\item{pvalue}{The name of the column containing p-values (default: 'pvalue').}

\item{padj}{The name of the column containing adjusted p-values (default: 'padj').}

\item{logFC}{The name of the column containing log2 fold changes (default: 'log2FoldChange').}

\item{gene_list}{A character vector of genes to label on the plot. If not empty,
only these genes will be labeled. If empty, the top and bottom genes
will be labeled based on significance (default: '').}

\item{padj_cutoff}{A numeric value for the adjusted p-value cutoff (default: 0.01).}

\item{logFC_cutoff}{A numeric vector of length two for the log2 fold change cutoffs.
The first value is for the lower bound (downregulation), and the second is for the
upper bound (upregulation) (default: c(-1, 1)).}

\item{basemean_cutoff}{A numeric value for the minimum mean expression level (baseMean)
to be considered for labeling in the default mode (default: 500).}

\item{top_n_genes}{An integer for the number of top-upregulated genes to label
when \code{gene_list} is empty (default: 10).}

\item{bottom_n_genes}{An integer for the number of top-downregulated genes to label
when \code{gene_list} is empty (default: 10).}
}
\value{
A ggplot object representing the volcano plot.
}
\description{
This function generates a volcano plot from a data frame,
highlighting genes with a significant adjusted p-value and log2 fold change.
It can also label a specific set of user-provided genes or the top/bottom
most significant genes based on log fold change.
}
\examples{
\dontrun{
# Assuming 'resLFC' is a differential expression results object

# Use the default labeling (top/bottom significant genes)
volcano_maker(df = resLFC \%>\% as_tibble(rownames = 'Gene'))

# Adjust all parameters for default labeling
volcano_maker(df = resLFC \%>\% as_tibble(rownames = 'Gene'),
              padj_cutoff = 0.05,
              logFC_cutoff = c(-2, 2),
              basemean_cutoff = 1000,
              top_n_genes = 5,
              bottom_n_genes = 5)

# Use a specific gene list for labeling
my_genes <- c("GeneA", "GeneB", "GeneC")
volcano_maker(df = resLFC \%>\% as_tibble(rownames = 'Gene'), gene_list = my_genes)
}
}
